requirements.txt
Flask==2.2.0
python-dotenv==0.19.2
pymongo==3.12.1
requests==2.28.1
python-telegram-bot==20.3
requests
gunicorn
